Add files via upload
Hello, this is a simple tool that I have prepared, that can analyze python Cryptocurrency. This tool can be developed and customized. You can easily get the averages of the current currency, such as the rise and fall, by changing it daily, monthly or the desired day length. It can be run as google colab or py extension.